backend/app/requirements.txt
fastapi
uvicorn
transformers
torch
soundfile
ffmpeg-python
python-multipart
